requirements.txt
Requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
camelot_py==0.11.0
pymupdf==1.25.1
opencv-python==4.10.0.84
numpy==2.2.0
pandas==2.2.3
psutil==6.1.0
Unidecode==1.3.8
selenium==4.27.1
webdriver_manager==4.0.2